[ success ] jointly query for referral

+++ db/src/entities/Position.ts
@@ -15,7 +15,7 @@ import { Referral } from "./Referral";
 @Entity("position", { schema: "aki" })
 export class Position {
     // Matt Cappucci - made this auto increment
-  @PrimaryGeneratedColumn({ type: "bigint", name: "id" })
+  @PrimaryGeneratedColumn({ type: "int", name: "id" })
   id: number;
 
   /*@Column("bigint", { primary: true, name: "id" })
